update blue focus ring for Input Toggle, Choice Set
- redesign new focus style
- update ACRTextView
- remove ACRHyperlinkTextview
