+++ source/macos/AdaptiveCards/AdaptiveCards/Views/ACRChoiceButton.swift
@@ -72,9 +72,9 @@ class ACRChoiceButton: NSView, NSTextFieldDelegate, InputHandlingViewProtocol {
         fatalError("init(coder:) has not been implemented")
     }
     
-     // Label
-    private (set) lazy var buttonLabelField: ACRHyperLinkTextView = {
-        let view = ACRHyperLinkTextView()
+    // Label
+    private (set) lazy var buttonLabelField: ACRTextView = {
+        let view = ACRTextView()
         view.translatesAutoresizingMaskIntoConstraints = false
         view.isEditable = false
         view.backgroundColor = .clear
@@ -270,7 +270,7 @@ extension ACRChoiceButton {
     var labelAttributedString: NSAttributedString {
         get { buttonLabelField.attributedString() }
         set {
-            buttonLabelField.textStorage?.setAttributedString(newValue)
+            buttonLabelField.setAttributedString(str: newValue)
         }
     }
     var state: NSControl.StateValue {
